pyproject.toml
[project]
name = "job-haunt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.10.1",
    "daphne>=4.2.1",
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "drf-spectacular>=0.28.0",
    "isort>=7.0.0",
    "pillow>=12.0.0",
]
